uW power tracking RX seems to work

+++ src/FSOcontroller.h
@@ -44,7 +44,7 @@ class FSOcontroller{
     private:
         void pollForPower();
         void exec();
-        std::vector<float> SFPpowers;
+        std::vector<int> SFPpowers;
         int highestPowerSFP = 1;
         int prevHighestPowerSFP = 1;
 #ifdef ROUTE_TX_ONLY_ONE_FIBRE
